# Improve circularbuffer performance by removing redundant fields